+++ Java/src/problem_modelings/budget_constrained_lmax_optimization/algorithms/SABudgetConstrainedLmaxOptimizationModeling.java
@@ -91,7 +91,7 @@ public class SABudgetConstrainedLmaxOptimizationModeling extends BudgetConstrain
         int maxL = super.calculateMaxL(modelPlainOldData.tempControllerXSpinVariables);
         double summationOfLMax = super.calculateDistanceToNearestControllerEnergy(modelPlainOldData.tempControllerXSpinVariables);
 
-        int reliabilityEnergy = Utils.getReliabilityEnergy(
+        double reliabilityEnergy = Utils.getReliabilityEnergy(
                 modelPlainOldData.graph,
                 modelPlainOldData.controllerY,
                 modelPlainOldData.candidateControllers, modelPlainOldData.tempControllerXSpinVariables,
